Move code related to Redis protocol to a separate file.
